+++ typescript/radix-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @hyperlane-xyz/radix-sdk
 
+## 30.1.1
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@30.1.1
+- @hyperlane-xyz/provider-sdk@4.2.5
+
+## 30.1.0
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@30.1.0
+- @hyperlane-xyz/provider-sdk@4.2.4
+
 ## 30.0.0
 
 ### Patch Changes